Fail closed deploy staging discovery
